[WCMSFEQ-1105] Convert AMD modules to ES6 modules (#273)
This represents a massive refactor of some 70-odd js files. The idea is to deprecate the use of require.js/AMD style modules (and UMD where unnecessary) in favor of ES6 modules (import/export). This is not a full refactor, but many singletons needed to be fixed and in a few other cases (eg a lingering global assignment) code was cleaned up minorly. In addition, the API for some imports changed and this is reflected in the changes to Common.js most clearly. A few files have not been converted, either because they are already deprecated and awaiting deletion or because they are jquery-ui extensions and would be a more substantial undertaking for less obvious gains. In addition Charts.js was maintained as an AMD module for the nonce.

* Refactor AMD modules into ES6 modulesIn some cases, the module API needed to be changed or simplified and sothe importing file was also edited.THe generated configuration.js file was changed as well, requiringa change to the grunt task.

* Fix execution order of some singletons

* Add release notes

* Update release notes

* add variable declairation to prevent eval error

* Convert jQuery ready event to DOMContentLoaded event

The jQuery ready event is executed in a seperate queue stack from DOMContentLoaded

* revert charts.js so it will continue to work with code in percussion

* add variable declarations

* Fix last 2 global declarations in analytics.js